refactor: Move all rpq function to rpq.py file
